Extract Range into sierra utils.
commit-id:8ba3d989
