+++ crates/cairo-lang-sierra-to-casm/src/invocations/range_reduction.rs
@@ -1,7 +1,8 @@
 use cairo_felt::Felt252;
 use cairo_lang_casm::builder::CasmBuilder;
 use cairo_lang_casm::casm_build_extend;
-use cairo_lang_sierra::extensions::range_reduction::{Range, RangeConcreteLibfunc};
+use cairo_lang_sierra::extensions::range_reduction::RangeConcreteLibfunc;
+use cairo_lang_sierra::extensions::utils::Range;
 use num_bigint::ToBigInt;
 use num_traits::Zero;
